LayoutTests/crypto/webkitSubtle/sha-512-expected.txt

CONSOLE MESSAGE: line 1: WebKitSubtleCrypto is deprecated. Please use SubtleCrypto instead.
Test crypto.webkitSubtle.digest.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


SHA512 of []
PASS bytesToHexString(new Uint8Array(digest)) is 'cf83e1357eefb8bdf1542850d66d8007d620e4050b5715dc83f4a921d36ce9ce47d0d13c5d85f2b0ff8318d2877eec2f63b931bd47417a81a538327af927da3e'
SHA512 of [0x0]
PASS bytesToHexString(new Uint8Array(digest)) is 'b8244d028981d693af7b456af8efa4cad63d282e19ff14942c246e50d9351d22704a802a71c3580b6370de4ceb293c324a8423342557d4e5c38438f0e36910ee'
SHA512 of [0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10]
PASS bytesToHexString(new Uint8Array(digest)) is '8ffaee0cccc162851faf051ae38667eefd423c0164c50055f8ade00afc3705e3cdeb9900004b0e426ca66ab63aa3b99b075273f44fd37c22a3555c6fd1f37ccb'
PASS successfullyParsed is true

TEST COMPLETE
